Move base64 stuff to v0
